requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
transformers==4.41.1
torch==2.2.2
nemo_toolkit[all]==1.23.0
omegaconf==2.3.0
pydub==0.25.1
python-multipart==0.0.9
